platform-profile: Work-around dytc_lapmode behaviour
The dytc_lapmode behaviour is pretty wonky. Instead of behaving like a
proximity sensor, it tells us whether it's actively changing the
performance profile, so it gets reset when something, whether the user
or more likely the daemon, changes the profile to balanced.

Work-around this by not changing the kernel platform_profile when the
performance mode gets inhibited, but advertise it as "balanced" to
end-users. The profile will get reset to "performance" when the
inhibition finishes, but only for hardware that uses the dytc_lapmode.

Whether or not this behaviour needs to apply to all drivers will need
design discussions.