utils: Add "tainting" detection
Throw a warning if custom fan curves are used on Asus systems as those
could potentially be setup in such a way as to not match intended
workloads, eg. it could make the performance profile have very little
cooling and throttling in a way that would be expected of a "quiet"
platform_profile.

See https://patchwork.kernel.org/project/platform-driver-x86/patch/20210820095726.14131-2-luke@ljones.dev/#24397127