Install custom debug message handler
Remove a lot of the unnecessary stuff about the process, and instead
pad out the domain to make it more accessible.

If running on an interactive terminal add color coding for warnings
and errors. If this behavior is not desirable for an interactive terminal,
NO_COLOR=0 can be set in the environment.