Fix running test suite without root
Running Tests.test_fake_driver breaks stdin on a real terminal unless
the test suite is run as root.  It doesn't make sense for the fake
driver to capture stdin unless we are sure that we're interactive.

Lookup whether stdout is a tty and only run the keyboard setup in
probe for the fake driver if we're sure we're interactive.